complete line drawing algo

+++ src/transform.c
@@ -6,7 +6,7 @@
 /*   By: rmei <rmei@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/09 15:53:58 by rmei              #+#    #+#             */
-/*   Updated: 2024/08/13 13:05:38 by rmei             ###   ########.fr       */
+/*   Updated: 2024/08/13 16:33:17 by rmei             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void    ft_iso_transform(t_vec_3D **vecs, int width, int height)
 {
     int         i;
     int         x;
-    int         a;
+    float         a;
     float       zoom;
     t_vec_3D    *v;
     
@@ -33,6 +33,7 @@ void    ft_iso_transform(t_vec_3D **vecs, int width, int height)
         x = v->x;
         v->x = (x * cos(a) - v->y * cos(a)) * zoom + I_WIDTH / 2;
         v->y = ((x * sin(a) + v->y * sin(a)) / 2 - v->z) * zoom + I_HEIGHT / 2;
+	    
         i++;
     }
 }
